Fifth commit

+++ src/Controller/LendingsController.php
@@ -13,22 +13,140 @@ use Cake\I18n\Time;
  */
 class LendingsController extends AppController
 {
+	public $paginate = [
+	'limit' => 100,
+	];
 
     /**
      * Index method
      *
      * @return \Cake\Http\Response|void
      */
+      public function pdfoindex($groupid)
+    {
+		$this->viewBuilder()->options([
+            'pdfConfig' => [
+                'orientation' => 'portrait',
+                'filename' => 'Llista de préstecs - ' . $groupid . '.pdf'
+            ]
+        ]);
+		
+		$plendings = $this->Lendings->find('all', [
+        'recursive'=>-1,
+        'fields' => [
+           'Lendings.id',
+           'Lendings.date_taken',
+           'Lendings.date_return',
+           'Lendings.date_real_return',
+           'lending_state_name' => 'LendingStates.name',
+           'Lendings.lending_state_id',
+           'document_id' => 'Documents.id',
+           'document_name' => 'Documents.name',
+           'set_lending_user_name' => 'SetLendingUsers.name',
+           'set_return_user_name' => 'SetReturnUsers.name',
+           'user_name' => 'Users.name',
+           'user_id' => 'Users.id',
+           'group_name' => 'Groups.name',
+           'group_id' => 'Groups.id',
+           'phone1' => 'Users.phone1',
+           'phone2' => 'Users.phone2',
+           'Lendings.created',
+           'Lendings.modified'
+         ],
+         'conditions' => ['Lendings.lending_state_id !=' => 5, 'Groups.id' => $groupid],
+        'order' => ['Users.name' => 'ASC'],
+        'contain' => ['Users' => ['Groups'], 'Documents', 'SetLendingUsers', 'SetReturnUsers', 'LendingStates'],
+        'sortWhitelist' => ['id', 'name', 'group_name', 'date_taken', 'date_return', 'lending_state_id', 'document_id', 'document_name', 'set_lending_user_name', 'set_return_user_name', 'user_name', 'created','modified']
+    ]);
+
+        $this->set(compact('plendings'));
+		
+    }
+     
     public function index()
     {
-        $this->paginate = [
-            'contain' => ['Documents', 'Users', 'SetLendingUsers', 'SetReturnUsers', 'LendingStates']
-        ];
-        $lendings = $this->paginate($this->Lendings);
+		$prevcurrurl = $this->request->here(true);
+		$currurl = str_replace('/', 'º', $prevcurrurl);
+		
+		 $where = [
+        'recursive'=>-1,
+        'fields' => [
+           'Lendings.id',
+           'Lendings.date_taken',
+           'Lendings.date_return',
+           'Lendings.date_real_return',
+           'lending_state_name' => 'LendingStates.name',
+           'Lendings.lending_state_id',
+           'document_id' => 'Documents.id',
+           'document_name' => 'Documents.name',
+           'set_lending_user_name' => 'SetLendingUsers.name',
+           'set_return_user_name' => 'SetReturnUsers.name',
+           'user_name' => 'Users.name',
+           'user_id' => 'Users.id',
+           'group_name' => 'Groups.name',
+           'group_id' => 'Groups.id',
+           'Lendings.created',
+           'Lendings.modified'
+         ],
+         'conditions' => ['Lendings.lending_state_id !=' => 5],
+        'order' => ['Documents.name' => 'ASC'],
+        'contain' => ['Users' => ['Groups'], 'Documents', 'SetLendingUsers', 'SetReturnUsers', 'LendingStates'],
+        'sortWhitelist' => ['id', 'name', 'group_name', 'date_taken', 'date_return', 'lending_state_id', 'document_id', 'document_name', 'set_lending_user_name', 'set_return_user_name', 'user_name', 'created','modified'],
+    ];
+    
+    
+     // Set pagination
+    $this->paginate = $where;
 
-        $this->set(compact('lendings'));
+    // Get data
+    $lendings = $this->paginate($this->Lendings, ['limit' => 100]);
+    $gr = $this->LoadModel('Groups');
+    $groups = $gr->find('list');
+
+        $this->set(compact('lendings', 'currurl', 'groups'));
+		
     }
+    
+    public function doneindex()
+    {
+		
+		$prevcurrurl = $this->request->here(true);
+		$currurl = str_replace('/', 'º', $prevcurrurl);
+		
+		 $where = [
+        'recursive'=>-1,
+        'fields' => [
+           'Lendings.id',
+           'Lendings.date_taken',
+           'Lendings.date_return',
+           'Lendings.date_real_return',
+           'lending_state_name' => 'LendingStates.name',
+           'Lendings.lending_state_id',
+           'document_id' => 'Documents.id',
+           'document_name' => 'Documents.name',
+           'set_lending_user_name' => 'SetLendingUsers.name',
+           'set_return_user_name' => 'SetReturnUsers.name',
+           'user_name' => 'Users.name',
+           'user_id' => 'Users.id',
+           'Lendings.created',
+           'Lendings.modified'
+         ],
+         'conditions' => ['Lendings.lending_state_id' => 5],
+        'order' => ['Documents.name' => 'ASC'],
+        'contain' => ['Users', 'Documents', 'SetLendingUsers', 'SetReturnUsers', 'LendingStates'],
+        'sortWhitelist' => ['id', 'name', 'date_taken', 'date_return', 'lending_state_id', 'document_id', 'document_name', 'set_lending_user_name', 'set_return_user_name', 'user_name', 'created','modified'],
+    ];
+    
+    
+     // Set pagination
+    $this->paginate = $where;
 
+    // Get data
+    $lendings = $this->paginate($this->Lendings, ['limit' => 100]);
+
+        $this->set(compact('lendings', 'currurl'));
+		
+    }
     /**
      * View method
      *
@@ -88,7 +206,84 @@ class LendingsController extends AppController
 	}
             
 		
-
+	public function quickadd($document_id, $user_id, $referer)
+	{
+		$documents = $this->loadModel('Documents');
+		$document = $documents->get($document_id);
+		if($document->lending_state_id == 1){
+		
+			$lending = $this->Lendings->newEntity();
+			if ($this->request->is('post')) {
+				$lending = $this->Lendings->patchEntity($lending, $this->request->getData());
+				$document = $this->Documents->patchEntity($document, $this->request->getData());
+				$lending->document_id = $document_id;
+						$lendinguser = $user_id;
+						$users = $this->loadModel('Users');
+						$user = $users->get($lendinguser, [
+							'contain' => ['Groups', 'Lendings']
+						]);
+						$items = 0;
+						if(!empty($user->lendings)){
+						foreach($user->lendings as $userlending){
+							if($userlending->lending_state_id == 2 || $userlending->lending_state_id == 2) { $items = $items + 1; }
+						}
+						}
+						$lendinggroup = $user->group_id;
+						$groups = $this->loadModel('Groups');
+						$group = $groups->get($lendinggroup, [
+							'contain' => ['LendingPolicies']
+						]);
+						$lendingpolicy_id = $group->lending_policy_id;
+						$lendingpolicies = $this->loadModel('LendingPolicies');
+						$lendingpolicy = $lendingpolicies->get($lendingpolicy_id);
+						$maxdays = $lendingpolicy->lending_max_days;
+						$maxitems = $lendingpolicy->lending_max_items;
+						
+						if($items == $maxitems){
+							$this->Lendings->delete($lending);
+							$this->Flash->error(__('Aquesta persona ja té '. $items . ' documents prestats, que és el màxim autoritzat per al seu grup.'));
+							$newreferer = $referer;
+							$ref = str_replace('º','/',$referer);
+							return $this->redirect('../../../../'.$ref);
+						}
+						
+						$now = Time::now();
+						$returndate = Time::now();
+						$returndate = $returndate->addDays($maxdays);
+						$lending->set_lending_user_id = $this->Auth->user('id');
+						$lending->date_taken = $now;
+						$lending->date_return = $returndate;
+						$lending->lending_state_id = 2;
+						$lending->user_id = $lendinguser;
+						$document->lending_state_id = 2;
+						$newreferer = $referer;
+						if ($this->Lendings->save($lending) && $this->Documents->save($document)) {
+							$this->Flash->success(__('S\'ha prestat el document a ' . $user->name . '. La data de retorn és: ' . $returndate->i18nFormat('d/M/YYYY') . '.'));
+							$ref = str_replace('º','/',$referer);
+							return $this->redirect('http://80.211.14.98/epergam2/users/view/' . $user->id);
+				}
+			
+            $this->Flash->error(__('The lending could not be saved. Please, try again.'));
+        }
+        }
+        elseif ($document->lending_state_id == 2){$this->Flash->error(__('Aquest document ja es troba en préstec.'));
+							$newreferer = $referer;
+							$ref = str_replace('º','/',$referer);
+							return $this->redirect('../../../../'.$ref);}
+        elseif ($document->lending_state_id == 3){$this->Flash->error(__('Aquest document ja es troba en préstec.'));
+							$newreferer = $referer;
+							$ref = str_replace('º','/',$referer);
+							return $this->redirect('../../../../'.$ref);}
+        elseif ($document->lending_state_id == 4){$this->Flash->error(__('Aquest document no es pot prestar perquè està reservat.'));
+							$newreferer = $referer;
+							$ref = str_replace('º','/',$referer);
+							return $this->redirect('../../../../'.$ref);}
+        elseif ($document->lending_state_id == 6){$this->Flash->error(__('Aquest document està exclòs de préstec.'));
+							$newreferer = $referer;
+							$ref = str_replace('º','/',$referer);
+							return $this->redirect('../../../../'.$ref);}
+        $this->set(compact('lending'));
+	}
     
     public function add($document_id, $referer)
     {
@@ -143,12 +338,12 @@ class LendingsController extends AppController
 						if ($this->Lendings->save($lending) && $this->Documents->save($document)) {
 							$this->Flash->success(__('S\'ha prestat el document a ' . $user->name . '. La data de retorn és: ' . $returndate->i18nFormat('d/M/YYYY') . '.'));
 							$ref = str_replace('º','/',$referer);
-							return $this->redirect('../../../../'.$ref);
+							return $this->redirect('http://80.211.14.98/epergam2/users/view/' . $user->id);
 				}
 			}
             $this->Flash->error(__('The lending could not be saved. Please, try again.'));
         }
-        $users = $this->Lendings->Users->find('list', ['limit' => 200]);
+        $users = $this->Lendings->Users->find('list');
         $this->set(compact('lending', 'users'));
     }
 
@@ -174,11 +369,11 @@ class LendingsController extends AppController
             }
             $this->Flash->error(__('The lending could not be saved. Please, try again.'));
         }
-        $documents = $this->Lendings->Documents->find('list', ['limit' => 200]);
-        $users = $this->Lendings->Users->find('list', ['limit' => 200]);
-        $setLendingUsers = $this->Lendings->SetLendingUsers->find('list', ['limit' => 200]);
-        $setReturnUsers = $this->Lendings->SetReturnUsers->find('list', ['limit' => 200]);
-        $lendingStates = $this->Lendings->LendingStates->find('list', ['limit' => 200]);
+        $documents = $this->Lendings->Documents->find('list');
+        $users = $this->Lendings->Users->find('list');
+        $setLendingUsers = $this->Lendings->SetLendingUsers->find('list');
+        $setReturnUsers = $this->Lendings->SetReturnUsers->find('list');
+        $lendingStates = $this->Lendings->LendingStates->find('list');
         $this->set(compact('lending', 'documents', 'users', 'setLendingUsers', 'setReturnUsers', 'lendingStates'));
     }
 
@@ -191,14 +386,21 @@ class LendingsController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $lending = $this->Lendings->get($id);
-        if ($this->Lendings->delete($lending)) {
-            $this->Flash->success(__('The lending has been deleted.'));
-        } else {
-            $this->Flash->error(__('The lending could not be deleted. Please, try again.'));
-        }
+		if($this->Auth->user('group_id') == 12){
+			$this->request->allowMethod(['post', 'delete']);
+			$lending = $this->Lendings->get($id);
+			if ($this->Lendings->delete($lending)) {
+				$this->Flash->success(__('The lending has been deleted.'));
+			} else {
+				$this->Flash->error(__('The lending could not be deleted. Please, try again.'));
+			}
 
-        return $this->redirect(['action' => 'index']);
-    }
+			return $this->redirect(['action' => 'index']);
+		}
+    else {
+		$this->Flash->error(__('No tens permisos per esborrar.'));
+		return $this->redirect(['controller' => 'Users', 'action' => 'login']);
+		
+		}
+}
 }
